lab09/SnakeGame/test/MyPanelTest.java
// For JUnit testing, focus on testing methods that can be tested independently of the GUI.
// This includes methods responsible for game logic and data manipulation, such as updating the snake's position,
// regenerating the apple, and checking for collisions.
// I'll write tests to ensure these methods behave as expected under various conditions.
import static org.junit.Assert.*;
import org.junit.Test;

import java.awt.event.KeyEvent;

public class MyPanelTest {
    // Initialize test panel and other necessary objects
    MyPanel panel = new MyPanel();

    @Test
    public void testKeyPressed() {
        // Test setting direction based on key presses
        panel.keyPressed(createKeyEvent(KeyEvent.VK_UP));
        assertEquals("Direction should be set to up", 1, panel.direction);

        panel.keyPressed(createKeyEvent(KeyEvent.VK_DOWN));
        assertEquals("Direction should be set to down", 2, panel.direction);

        panel.keyPressed(createKeyEvent(KeyEvent.VK_LEFT));
        assertEquals("Direction should be set to left", 3, panel.direction);

        panel.keyPressed(createKeyEvent(KeyEvent.VK_RIGHT));
        assertEquals("Direction should be set to right", 4, panel.direction);
    }

    @Test
    public void testRegenApple() {
        // Test if apple is regenerated within panel boundaries
        panel.regenApple();
        assertTrue("Apple x-coordinate should be within panel width", panel.apple_loc.x < 400);
        assertTrue("Apple y-coordinate should be within panel height", panel.apple_loc.y < 400);
    }

    // Helper method to create KeyEvent objects
    private KeyEvent createKeyEvent(int keyCode) {
        return new KeyEvent(panel, KeyEvent.KEY_PRESSED, System.currentTimeMillis(), 0, keyCode, KeyEvent.CHAR_UNDEFINED);
    }
}

//Future Improvement:s
//Modularization: Refactor the code to separate concerns and adhere to a more structured design pattern like MVC (Model-View-Controller) to improve maintainability and scalability.
//Enhanced Gameplay: Add features such as increasing speed over time, adding obstacles, or power-ups to make the game more challenging and engaging.
//User Interaction: Implement functionality for pausing/resuming the game, restarting the game, or displaying the score to enhance user experience.